Mod External reference resolver should not overwrite reference links on save
Utilize a "fail-through" strategy while ensuring that saving does not overwrite any existing reference link.

Utilize a before create event handler to override the create method.
This new create will prevent the save if the reference link already exists.
The existing reference link will be returned, mocking the save process, so that the caller need not care whether or not the reference link already existed before saving.

Alternative (and perhaps preferred) methods, such as those described in the link below, were not function as expected:
- https://docs.spring.io/spring-data/rest/docs/current/reference/html/#customizing-sdr.overriding-sdr-response-handlers